Move ResourceLoader out of renderer/
